refactor: presentation 패키지 구조 internal/external로 분리
