+++ src/main/java/com/yeoljeong/tripmate/presentation/external/PlanController.java
@@ -1,4 +1,4 @@
-package com.yeoljeong.tripmate.presentation;
+package com.yeoljeong.tripmate.presentation.external;
 
 import com.yeoljeong.tripmate.application.dto.command.ParticipatePlanCommand;
 import com.yeoljeong.tripmate.application.dto.result.ConfirmPlanUnitResult;
